% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abline_range.R
\name{abline_range}
\alias{abline_range}
\title{Add a line to a plot}
\usage{
abline_range(a = NULL, b = NULL, reg = NULL, from = NULL, to = NULL,
  ...)
}
\arguments{
\item{a}{Intercept (optional)}

\item{b}{Slope (optional)}

\item{reg}{A fitted linear regression model (output of \code{\link{lm}}).}

\item{from}{Draw from this X value}

\item{to}{Draw to this x value}

\item{\dots}{Further parameters passed to \code{\link{segments}}}
}
\description{
As \code{abline}, but with \code{from} and \code{to} arguments.
If a fitted linear regression model is used as asn argument, it uses the min and max values of the data used to fit the model.
}

