% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loessList.R
\name{loessList}
\alias{loessList}
\title{List of loess models}
\usage{
loessList(model, data, ...)
}
\arguments{
\item{model}{Formula of the form y ~ x|group}

\item{data}{A dataframe containing the three variables}

\item{\dots}{Further arguments passed to \code{\link{loess}}, such as \code{span} and \code{degree}.}
}
\description{
Like \code{\link{nlsList}}, but for \code{\link{loess}}. Supports only a single grouping variable, with a formula of the form "y ~ x|group".
}
\examples{
data <- as.data.frame(Orthodont)
data$age <- jitter(data$age, 2)
res <- loessList(distance ~ age | Sex, data=data)
}
