% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_regres_line.R
\name{add_regres_line}
\alias{add_regres_line}
\title{Add a regression line and confidence interval to a plot}
\usage{
add_regres_line(fit, from = NULL, to = NULL, band = TRUE,
  bandcolor = alpha("lightgrey", 0.8), ...)
}
\description{
Add a regression line and confidence interval to a plot
}
\examples{
fit <- lm(height ~ age, data=Loblolly)
with(Loblolly, plot(age, height, panel.first=add_regres_line(fit)))
}

