% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nls.R
\name{plot_nls}
\alias{plot_loess}
\alias{plot_nls}
\title{Plot a non-linear or non-parametric regression model}
\usage{
plot_nls(object, col = NULL, lines.col = palette(),
  points.col = palette(), ci.col = "#BEBEBEB3", lwd = 1, lty = 1,
  add = FALSE, xlab = NULL, ylab = NULL, coverage = 0.95, ...)

plot_loess(object, ...)
}
\description{
Convenient function for adding curves to an existing plot, or to plot the data used in fitting with the curve overlaid. Works for simple non-linear regression models fit with \code{\link{nls}}, and grouped non-linear regression (with \code{\link{nlsList}}).
}
\examples{

chick <- as.data.frame(ChickWeight)
fit0 <- nls(weight ~ a*Time^b, data=chick, start=list(a=10, b=1.1))
plot_nls(fit0)

fit1 <- nlsList(weight ~ a*Time^b|Diet, data=chick, start=list(a=10, b=1.1))
plot_nlslist(fit1)

}

