% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nls.R
\name{plot_nls}
\alias{plot_nls}
\title{Plot a non-linear regression model}
\usage{
plot_nls(object, lines.col = palette(), points.col = palette(),
  add = FALSE, ...)
}
\description{
Convenient function for adding curves to an existing plot, or to plot the data used in fitting with the curve overlaid. Works for simple non-linear regression models fit with \code{\link{nls}}, and grouped non-linear regression (with \code{\link{nlsList}}).
}
\examples{

chick <- as.data.frame(ChickWeight)
fit0 <- nls(weight ~ a*Time^b, data=chick, start=list(a=10, b=1.1))
fit1 <- nlsList(weight ~ a*Time^b|Diet, data=chick, start=list(a=10, b=1.1))

plot_nls(fit0)
}

